+++ client_it_test.go
@@ -1683,7 +1683,7 @@ func TestRowIDColumnIDImportRoaring(t *testing.T) {
 func TestRowIDColumnIDTimestampImportRoaring(t *testing.T) {
 	client := getClient()
 	iterator := newTestIteratorWithTimestamp()
-	field := index.Field("importfield-rowid-colid-time", OptFieldTypeTime(TimeQuantumMonthDayHour))
+	field := index.Field("importfield-rowid-colid-time", OptFieldTypeTime(TimeQuantumYearMonthDayHour))
 	err := client.EnsureField(field)
 	if err != nil {
 		t.Fatal(err)